+++ lib/src/logic_mixins/page_list_handler.dart
@@ -33,6 +33,7 @@ mixin PageListHandler {
             page: page,
             pdfWorker: pdfWorker,
             controller: tPdfController,
+            readerStateController: stateController,
           ),
         ),
       );