+++ lib/t_pdf_reader.dart
@@ -1,2 +1,6 @@
 export 'src/t_pdf_reader_base.dart';
+export 'src/default_widgets/scrollbar_widgets.dart';
+export 'src/events/pdf_events.dart';
+export 'src/events/user_events.dart';
+
 export 'package:than_pdf_engine/core/pdf_core.dart';