+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.0
+
+- Added Event `Keyboard` -> `KeyDown`,`KeyUp`
+- Added `TPdfController:keyboardScrollSpeed` - Property.
+- Added `TPdfController:onKeyEvent` - Function.
+- Changed `mixin` -> `Interface Component` Logic.
+- Fixed `Worng Pdf Page Count`
+
 ## 0.1.0
 
 - Update `than_pdf_engine`